Improve liboath usersfile write handling.
Thanks to Fabian Vogt of SUSE for finding and reporting the problem.

Open *.lock and *.new file in exclusive mode to treat any existing
files as an error, including if it is a symbolic link.

Modify so that if opening the lock file fails, it returns
OATH_FILE_LOCK_ERROR instead of OATH_FILE_CREATE_ERROR so that the
latter is reserved to mean error to create the target file.

Fix bug that causes stale *.lock file to be kept around on failure to
open *.new file.